+++ src/main/java/org/monarchinitiative/phenotefx/service/Resources.java
@@ -20,13 +20,16 @@ package org.monarchinitiative.phenotefx.service;
  * #L%
  */
 
+import ontology_term.BiologySex;
 import org.monarchinitiative.phenol.ontology.data.Ontology;
 import org.monarchinitiative.phenotefx.io.EctoParser;
 import org.monarchinitiative.phenotefx.io.HPOParser;
 import org.monarchinitiative.phenotefx.io.MedGenParser;
 import org.monarchinitiative.phenotefx.io.MondoParser;
 
+import java.util.HashMap;
 import java.util.Map;
+import java.util.stream.Collectors;
 
 
 /**