+++ CHANGELOG.md
@@ -30,7 +30,7 @@
 * fixed accelerator bug on linux
 
 ## v0.3.3
-* Added field for clinical modifier, updated for phenol 
+* Added field for clinical effectSizeType, updated for phenol 
 
 ## v0.3.1
 * Refactored PhenoteFX to use the new V2 small files.